requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai>=2.0.0
PyPDF2==3.0.1
python-dotenv>=1.0.0
